examples/python-core-example/requirements.txt
requests==2.23.0
urllib3==1.25.7
retry==0.9.2
pyyaml==5.4.1
pytz==2021.1
munch==2.5.0
click==8.0.3
tqdm==4.62.3